requirements-s1.txt
matplotlib==3.5.0
sentencepiece==0.2.0
Pillow==9.5.0
pyopengl
PyOpenGL_accelerate
PyYAML>=6.0
yacs>=0.1.8
scikit-image==0.19.1
termcolor
tqdm==4.62.3
loguru==0.5.3
trimesh>=3.17.1
flatten_dict==0.4.2
ipykernel==5.3.4
ipywidgets==7.6.5
jpeg4py
shapely==1.7.1
vedo==2020.4.2
rtree==0.9.7
gdown==4.6.0
pytorch_lightning==1.2.5
PyMCubes==0.1.4
kornia
chumpy
tinyobjloader==2.0.0rc7
cython==0.29.20
opencv-python
opencv_contrib_python
scikit-learn==1.3.2
scipy==1.10.1
simple-romp==1.0.4
protobuf==3.20.0
open3d
dataclasses>=0.6
numpy==1.24.4
torch==1.13.0 
torchvision
omegaconf
smplx[all]
diffusers==0.31.0
accelerate==1.0.1
transformers==4.46.3
safetensors==0.4.5
huggingface_hub
git+https://github.com/facebookresearch/pytorch3d.git@v0.7.1
git+https://github.com/YuliangXiu/neural_voxelization_layer.git
git+https://github.com/NVIDIAGameWorks/kaolin.git
git+https://github.com/YuliangXiu/rembg.git
git+https://github.com/nghorbani/human_body_prior.git
git+https://github.com/nghorbani/body_visualizer.git
git+https://github.com/openai/CLIP.git
